requirements.txt
apify-client==1.7.1
python-dotenv==1.0.0
pydantic==2.5.0
colorlog==6.8.0
colorama==0.4.6
dnspython==2.4.2
pymongo==4.6.1